pyproject.toml
[tool.poetry]
name = "photo-renamer"
version = "0.1.0"
description = ""
authors = ["Filip Šustík <filip.sustik@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7 <3.11"
python-dotenv = "^0.19.2"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"
pyinstaller = "^4.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
photo-renamer = 'photo_renamer.__main__:main'

[tool.isort]
profile = "black"
multi_line_output = 3
known_first_party = "photo_renamer"

[tool.pytest.ini_options]
testpaths = [ "photo_renamer", "tests" ]
addopts = "--pyargs --doctest-modules"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true